fix(automatic): ensure `window.location` is defined
* Replaced custom input logic with reusable SetpointInput component
* Introduced mount-time check
    *   ensure `window.location` is accessed only on the client

issue #29
